move parseQS to utils
