+++ src/config.js
@@ -16,13 +16,12 @@ import { isValidPriceConfig } from './cpmBucketManager.js';
 import find from 'core-js/library/fn/array/find.js';
 import includes from 'core-js/library/fn/array/includes.js';
 import Set from 'core-js/library/fn/set.js';
-import { parseQS } from './url.js';
 
 const from = require('core-js/library/fn/array/from.js');
 const utils = require('./utils.js');
 const CONSTANTS = require('./constants.json');
 
-const DEFAULT_DEBUG = (parseQS(window.location.search)[CONSTANTS.DEBUG_MODE] || '').toUpperCase() === 'TRUE';
+const DEFAULT_DEBUG = (utils.parseQS(window.location.search)[CONSTANTS.DEBUG_MODE] || '').toUpperCase() === 'TRUE';
 const DEFAULT_BIDDER_TIMEOUT = 3000;
 const DEFAULT_PUBLISHER_DOMAIN = window.location.origin;
 const DEFAULT_ENABLE_SEND_ALL_BIDS = true;
