+++ src/url.js
@@ -1,19 +1,4 @@
-export function parseQS(query) {
-  return !query ? {} : query
-    .replace(/^\?/, '')
-    .split('&')
-    .reduce((acc, criteria) => {
-      let [k, v] = criteria.split('=');
-      if (/\[\]$/.test(k)) {
-        k = k.replace('[]', '');
-        acc[k] = acc[k] || [];
-        acc[k].push(v);
-      } else {
-        acc[k] = v || '';
-      }
-      return acc;
-    }, {});
-}
+import { parseQS } from './utils.js';
 
 export function formatQS(query) {
   return Object
